pyproject.toml
[tool.poetry]
name = "cookiecutter-hypermodern-python-ml"
version = "0.1.0"
description = "Project template for deployable ML/AI applications"
authors = ["Matthew Sach <78856091+mdsach@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
nox-poetry = "^1.0.1"
pre-commit = "^2.20.0"
github3-py = "^3.2.0"
click = "^8.1.3"
isort = "^5.10.1"
sphinx = "^5.3.0"
sphinx-autobuild = "^2021.3.14"
myst-parser = "^0.18.1"
furo = "^2022.9.29"
tomli = "^2.0.1"
pre-commit-hooks = "^4.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
